refactor : SearchDetailScreen 패키지 이동
